Reorganize/split code in Assignment.h/.cpp
